modules/SearchService/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opensearch-py==2.4.2
python-dotenv==1.0.0
pydantic==2.5.2
